Handle multiple highlighted words per section, and per line.
It can be that Diogenes outputs two highlighted words between two hr
elements. Previously we assumed that there could be only one, and only
considered the final one (using, however, the "context" link for the
first one). Now we output each one, using the "context" link for the
first one, and incrementing line numbers for the lines following.

It still can happen that two sections overlap; e.g., lines A and B
appear in one section, and the same lines A and B appear in the
following section. In the first section, the "context" link refers to
line A, while in the second, it refers to line B. We'll have incorrect
output in this case. If the line number of line A is n, and B comes 5
lines later, we'll output
	line n		A
	line n+5	B
	line n+5	A
	line n+5+5	B