Correctly initialize weighted/unweighted based on state of the radio buttons.
The bug here was having the "change" event examine the `value` of the
radio button itself, not of the RadioNodeList containing it. The `value`
of a particular radio button never changes, whether it is checked or
unchecked; only the `value` of the RadioNodeList changes. The display
was always initializing to unweighted, simply because the unweighted
radio button was the last to have its synthetic "change" event
dispatched.

Fixes #22.