Have TEI.lines emit a sequence of WORD/NONWORD tokens.
Instead of a text string. This will help to distinguish U+2019 used as a
right single quotation mark as can be used to render the end of a <q>
element (NONWORD), and U+2019 as an apostrophe used as a mark of elision
that belongs to the previous word (WORD).

This also makes the HTML output slightly more semantic, as following
nonword punctuation is no longer included in a sedes <span>, unless it
is between two words with the same sedes.