Restore the meaning of `default` in lemma.lookup.
I would expect BackoffGreekLemmatizer to return a lemma of None when it
exhausts its chain of
	DictLemmatizer
	UnigramLemmatizer
	RegexpLemmatizer
	DictLemmatizer
	IdentityLemmatizer
But there seems to be a bug: the UnigramLemmatizer returns not None, but "",
for διοτρεφές, which short-circuits the loop over self._taggers in
cltk/lemmatize/backoff.py SequentialBackoffLemmatizer.tag_one. Anyway,
we're now catching that "" and replacing it with the default.