Let expectancy CSV files contain only (lemma, sedes) ⇒ score mapping.
Formerly the expectancy file was the same as the output of tei2csv, only
with extra columns. This resulted in a lot of redundancy, as the same
scores were stored in the file for every occurrence of a lemma/sedes
pair. Because the only input to the expectancy calculation is lemma and
sedes, we only need to store a score once per unique lemma/sedes pair.

With this we also rework how tei2html works. Formerly, it looked up the
original words in the expectancy CSV using a key of
work/book_n/line_n/word_n. This had the problem that the keys were not
always unique, because line numbers are sometimes duplicated. Now we
re-do the same splitting/scansion/lemmatization of tei2csv in tei2html,
and look up the expectancy score in the expectancy CSV file. This has
the side benefit that the text tei2html processes does not necessarily
have to have been part of the input to the expectancy CSV file. Because
tei2html no longer does the reverse lookup, it no longer takes a work
identifier on its command line.