Set `color-adjust: exact`.
This causes the colors to remain when the page is printed. Without this,
the background colors making up the color scale disappear, and the
colors on text become more muted (perhaps having been converted into
CMYK from sRGB).

This works the way I expect it to in Chromium 88. It works only
partially in Firefox 78—the background colors work but the text colors
become black, except for the most saturated red and blue. It is supposed
to be working in Firefox 81:
https://bugzilla.mozilla.org/show_bug.cgi?id=1436758#c11