Canonicalize the order of diacritics from Beta Code.
When we are attaching Unicode combining characters to a base character,
we need to take care with certain diacritics so that they normalize into
NFC nicely. The basic rule is that we want to sort these three
diacritics:
  ( ) +
before these three:
  / \ =
Getting the order wrong prevents full NFC composition. For example,
`i(/`, after decoding and normalizing, becomes the single character
  U+1f35 GREEK SMALL LETTER IOTA WITH DASIA AND OXIA
But `i/(`, if the diacritics were kept in that order, normalizes only
partially:
  U+03af GREEK SMALL LETTER IOTA WITH TONOS
  U+0314 COMBINING REVERSED COMMA ABOVE

This change necessitated the reordering of some combining characters in
known.py, to make known-scansion lookups keep working.

Closes #45.