Ignore identity lemmata when trying word transformations.
The algorithm for trying transformations tries each one in turn,
stopping at the first one for which the CLTK backoff lemmatizer returns
a lemma. I neglected to consider that the CLTK backoff lemmatizer itself
has a last-resort identity lemmatization step, so it virtually always
returns a lemma (the only exceptions are words that are affected by
https://github.com/cltk/cltk/issues/1051). This meant that in almost all
cases, only the first (identity) transformation was tried, and the
grave→acute and acute→null transformations were never even considered.

Now we activate verbose mode on the CLTK backoff lemmatizer, which tells
us which sub-lemmatizer is responsible for the returned lemma, and
ignore those that come from IdentityLemmatizer.