Omit fallback identity lemmata in tei2csv output.
This allows us to distinguish words which actually are their own lemma,
from those for which lemma lookup failed. As identity lemmata are
trivial to compute, we can always swap them in at the point of need, in
expectancy or tei2html, rather than encoding them into the input. Now it
is possible to get a list of words without lemmata, by processing the
tei2csv CSV files.

Closes #48.