Downgrade Cardo roman from 1.04 to 0.99.
1.04 has some weird line-height issue that misaligns the shade bubbles
and causes the "Highlight undefined z-scores" highlight to cover only
the upper half of the characters. Only the upper half becomes a click
target as well. Furthermore, the font was causing this warning in the
Firefox console:
	downloadable font: hhea: bad linegap: -399 (font-family: "Cardo" style:normal weight:400 stretch:100 src index:2) source: file://sedes/web-demo/fonts/Cardo-Regular.woff

The upstream link http://scholarsfonts.net/cardo99.zip is dead, but I
got an old copy from the Wayback Machine:
https://web.archive.org/web/20101128095824/http://scholarsfonts.net/cardo99.zip

The new Cardo99s.ttf from cardo99.zip from the Wayback Machine is
identical to the independently archived
https://github.com/ryanfb/Cardo/blob/v0.99/Cardo.ttf

I downloaded the contemporaneous italic from
https://web.archive.org/web/20120124123526/http://scholarsfonts.net/cardoita.zip
(The current link http://scholarsfonts.net/cardoita.zip is also dead.)
cardoita.zip happens to contain a Cardoi99.ttf that is identical to the
one that was in cardo104.zip, so this change does not affect the italic
at all.

We no longer have a bold version of the font. The bold was new in 1.04.

Resolves #56.