Direction-aware trim of tokens in tei.trim_tokens.
Only trim whitespace from the left of left-side tokens, and from the
right of right-side tokens. This can matter when, for example, a line
ends with a space and an em dash " —". The space should be retained;
i.e., the token should remain " —", not be trimmed to "—".

This change only affects a few files, and in only a few lines:
callimachushymns	2
homerichymns		2
nonnusdionysiaca	11
quintussmyrnaeus	1