Don't emit duplicate HTML ids.
We keep a cache of previously seen book and line ids. When an id for a
line has already been seen, we not only omit the id on the line's
element, but we omit the id on every word of that line. (As opposed to
just inserting the word ids into the cache.) This is because we don't
want to id only the last few words in a line with a duplicate line
number, if it happens to contain more words than an already emitted
line; e.g.:

	line 15		id="line-15"
	line 15 word 1	id="line-15-word-1"
	line 15 word 2	id="line-15-word-2"
	line 15 word 3	id="line-15-word-3"
	...intervening lines...
	line 15		id=None
	line 15 word 1	id=None
	line 15 word 2	id=None
	line 15 word 3	id=None
	line 15 word 4	id="line-15-word-4"
	line 15 word 5	id="line-15-word-5"

Closes #30.