Add a --by option to expectancy.
This allows controlling the "dist vars" and "cond vars" according to
which expectancy is computed. The convention is to list dist_vars first
and cond_vars second, separated by a slash. The default is
"sedes/lemma". The notation "A/B" is meant to resemble the | of
conditional probability: Pr(A|B).