Make lemma+sedes internal parameters in tei2html.
As in the expectancy program, this controls the "dist" and "cond" vars
over which expectancy is computed.

The JavaScript EXPECTANCY is now a two-level Map, with the first level
keyed by cond_vars and the second level keyed by dist_vars, instead of
having everything at one level keyed by `${lemma}/${sedes}`. This is
because, if dist_vars is permitted to range over variables other than
sedes, we may not have a convenient list of every possible value, as we
have with POSSIBLE_SEDES. Instead, quantities like sum_x are computed by
summing all the entries in the dist_vars map after looking up cond_vars
in the top-level map.

The sedes distribution in the infobox poses a problem when variables
other than sedes may be used. We might consider making it fully
generalized, with one column for every possible or attested value of
dist_vars. But with some variables, that might mean thousands of columns
in the table. Instead, we'll leave the table specialized with columns
for sedes, but still let it generalize for other variables. In each
column of the table, we'll show the x and z values for the currently
selected word, as if we had replaced that word's sedes with the sedes of
the column.