Have tei2csv output the original word if lemmatization failed.
Formerly we left this as an empty string in tei2csv output, as a signal
that lemmatization failed and the lemma should be inferred to be the
same as the word. But the main effect of that was just adding a
requirement to do such an inference in the downstream programs
expectancy and join-expectancy.