Undo an erroneous square bracket replacement in homerichymns.xml 4.81a.
`[1` and `]1` are Beta Code for left and right parenthesis, not literal
square brackets.

But the book text has square brackets here, so the parentheses were
probably a mistake and should have been square brackets in the first
place.
https://archive.org/details/hesiodhomerichym00hesi_0/page/330/mode/1up

So the fix here just to remove the `1s` that were left behind earlier.