Do lemma overrides as a separate step, rather than use DictLemmatizer with backoff.
This makes a difference only in one specific case: when the lemma for an
override is an empty string. This happens only for the partial word "π"
in lemma-overrides.csv. In that case, the backoff CLTK lemmatizer would
assume that no lemma had been found, and continue with its own
algorithm, ultimately lemmatizing "π" as "ἀπό".