Allow specifying (work, book_n, line_n, word_n) in lemma overrides.
If a row in lemma-overrides.csv has a work, book_n, line_n, word_n
specified, it applies only to that specific instance of the word. If
work, book_n, line_n, word_n are blank, then it applies to any instance
of the word (that isn't one of the more specific instances).

Closes #91.