Readjustments to exceptional-appositives.csv.
For corpus changes that were made since epic-rhythm.

10.15–10.20
15.58
https://github.com/sasansom/sedes/commit/79d1295ff10cdf853801328b597ac72bc6ccfa49

14.2
This one was not a corpus change since epic-rhythm, but rather it
appears to be an effect of continuous word_n numbering. 14.2 is split
into typographical lines of 1 + 2 + 6 words. What had been counted as
the 5th word in the third typographical line is now counted as the 8th
word from the beginning of the metrical line.