Permit zero-quantity non-vocalic words at sedes 13.
Elsewhere in a line, a zero-quantity word takes on the same sedes as the
words that follow it. When such a word is at the end of the line, the
only reasonable option is to put it at sedes 13.

There are no instances in the corpus where this happens, a zero-quantity
word at the end of a line. It could theoretically happen, I guess. The
real reason for making this change is the analysis of a line that that
is likely currently erroneous, Q.S. 12.38:

<lb rend="displayNum" n="38" />*palla/di xwome/nh| *trw/wn u(/per ai)xmhta/wn:'

The source line is wrong. See, at the end, that there is a punctuation
mark (middle dot, Beta code ':') followed by an apostrophe. This line is
analyzed by tei2csv as:

work,book_n,line_n,word_n,word,lemma,sedes,metrical_shape
Q.S.,12,38,1,παλλάδι,παλλάς,1,–⏑⏑
Q.S.,12,38,2,χωομένῃ,χάω,3,–⏑⏑–
Q.S.,12,38,3,τρώων,τρώς,6,––
Q.S.,12,38,4,ὕπερ,ὑπέρ,8,⏑⏑
Q.S.,12,38,5,αἰχμητάων’,αἰχμητάων’,9,––––

The punctuation mark disappears and is the apostrophe ends up joined to
the preceding word, which is incorrect, even given the error in the
source line. I am about to commit a change that makes the final
apostrophe instead be analyzed as a separate word, which is still
incorrect generally, but is more correct given the error in the source
line. This will make the apostrophe a final "word" by itself at
sedes 13.