Process the TEI surplus element.
https://tei-c.org/release/doc/tei-p5-doc/en/html/ref-surplus.html
https://tei-c.org/release/doc/tei-p5-doc/en/html/PH.html#PHOM

Not sure of the best way to handle it. The element appears in Hom.Hymn 18:
https://github.com/sasansom/canonical-greekLit/blob/bffbfd51323caf6dce0f288676c1919832775e19/data/tlg0013/tlg018/tlg0013.tlg018.perseus-grc2.xml#L110
<l n="12"><milestone unit="Para" ed="P"/>[<surplus resp="editor" reason="alternative to ll. 10-11">χαῖρ᾽. Ἑρμῆ χαριδῶτα, διάκτορε, δῶτορ ἐάων.</surplus>]</l>

It shows with brackets and a footnote in the book scan:
https://archive.org/details/hesiodhomerichym00hesiuoft/hesiodhomerichym00hesiuoft/page/442/mode/1up

If we were to remove the surplus, we might want to restructure it so
that it doesn't end up appearing as a line that contains only empty
brackets. Put the <l> inside the <surplus>, for example.