Remove known scansion of Q.S. 13.408.
This was actually being used to make a textual correction, the addition
of the word δή. That's better done by fixing the source texts. In fact
this line 13.408 had been fixed to add the δή in the master branch, in
f7de15cc7ec4961e918ed5d646899ef8e9a81443, so the known scansion lookup
didn't even have an effect there. Here, with the Perseus 6 texts that
don't have the corresponding fix yet, the known scansion lookup *was*
working, and causing an error in tei2html because of the addition of a
scansion word that's not in the TEI.