Refactor TEI parsing.
What had been done in the TEI.lines method is now divided into three
layers: one to parse the XML and emit a sequence of low-level "events",
one to adjust the sequence of events to better match the semantics we
are aiming for, and one to bundle up the rectified events into Lines and
Tokens as before.

The computational nesting is flatter and the state more local. This
refactoring is done with the aim of supporting the rend="merge"
convention for joining quotations across multiple lines.