Trim whitespace from l and q elements.
We don't want whitespace at the beginning and end of lines, even though
there sometimes is in the source TEI. But trim_tokens was already taking
care of that. The bigger reason for this change is that whitespace gets
in the way of merging quotations with `<q rend="merge">`. When we find
`<q rend="merge">`, we do a sanity check to make sure that what comes
immediately before it is another q element. (Possibly with intervening
LINE_BEGIN or LINE_END: quotes may be merged across line breaks.) The
sanity check is to catch things like this:
	<q>abc</q> XYZ <q rend="merge">def</q>
It's at best ambiguous what this means: is ` XYZ ` meant to be included
in the merged quotation? We raise an error to be safe. But there are
cases in Perseus that are, because of whitespace, formally the same,
such as at Dion. 40.545–546:
	<l xml:base="urn:cts:greekLit:tlg2045.tlg001.perseus-grc2:40" n="545"><q rend="merge"><q rend="merge">ἀλλʼ ἐρέεις· <q>μεθέπω διερὸν γένος, ἐκ δὲ ῥοάων</q> </q></q></l>
	<l xml:base="urn:cts:greekLit:tlg2045.tlg001.perseus-grc2:40" n="546"><q rend="merge"><q rend="merge"><q rend="merge">αὐτοτελὴς γενόμην, καὶ ἐμὴ τροφὸς ἔπλετο πηγή.</q></q></q></l>
Notice the `</q> </q></q>` in the first line. The first two
`<q rend="merge">` in the second line match the last two `</q>` in the
first line. But the third `<q rend="merge">` technically has a space
between it and the `</q>` it is supposed to be matched to.

Arguably cases like these are bugs in the TEI that should simply be
modified. But I'll try this approach of trimming the whitespace to see
if it gives satisfactory results without having to push a lot of
whitespace changes upstream to Perseus, or maintain them ourselves.